More tests for adding Beacons/Instances
