+++ beacons/beacons_test.go
@@ -17,66 +17,13 @@ package beacons
 import (
     "testing"
 
-    "fmt"
+    "bytes"
+    "net/http"
+    "net/http/httptest"
 
+    "github.com/gorilla/mux"
     "github.com/stretchr/testify/assert"
-
-    "github.com/lighthouse/lighthouse/beacons/aliases"
 )
-
-func setup() (func()) {
-    SetupTestingTable()
-    aliases.SetupTestingTable()
-
-    return func() {
-        TeardownTestingTable()
-        aliases.TeardownTestingTable()
-    }
-}
-
-func Test_AddBeaconData(t *testing.T) {
-    teardown := setup()
-    defer teardown()
-
-    users := userMap{"USER":true}
-
-    testBeaconData := beaconData{
-        "BEACON_ADDR", "TOKEN", users,
-    }
-
-    addBeacon(testBeaconData)
-
-    var values beaconData
-    beacons.SelectRowSchema(nil, nil, &values)
-
-    assert.Equal(t, testBeaconData, values)
-}
-
-func Test_UpdateBeaconData(t *testing.T) {
-    teardown := setup()
-    defer teardown()
-
-    testBeaconData := map[string]interface{}{ 
-        "Address" : "BEACON_ADDR_FAIL", 
-        "Token" : "TOKEN_FAIL", 
-        "Users" : userMap{"USER_FAIL":true},
-    }
-
-    beacons.InsertSchema(testBeaconData)
-
-    userPass := userMap{"USER_PASS":true}
-    keyData := beaconData {
-        "BEACON_ADDR_PASS", "TOKEN_PASS", userPass,
-    }
- 
-    updateBeaconField("Token", "TOKEN_PASS", "BEACON_ADDR_FAIL")
-    updateBeaconField("Users", userPass, "BEACON_ADDR_FAIL")
-    updateBeaconField("Address", "BEACON_ADDR_PASS", "BEACON_ADDR_FAIL")
-
-    values := beaconData{"BEACON_ADDR_FAIL", "TOKEN_FAIL", userMap{}}
-    beacons.SelectRowSchema(nil, nil, &values)
-    assert.Equal(t, keyData, values)
-}
 
 func Test_GetBeaconAddress_Found(t *testing.T) {
     teardown := setup()
@@ -109,41 +56,6 @@ func Test_GetBeaconAddress_NotFound(t *testing.T) {
 
     assert.Equal(t, "", res, 
         "GetBeaconAddress should give empty string on error")
-}
-
-func Test_GetBeaconData_Found(t *testing.T) {
-    teardown := setup()
-    defer teardown()
-
-    users := userMap{"USER":true}
-
-    testBeaconData := map[string]interface{}{
-        "Address" : "BEACON_ADDR", 
-        "Token" : "TOKEN", 
-        "Users" : users,
-    }
-
-    beacons.InsertSchema(testBeaconData)
-
-    res, err := getBeaconData("BEACON_ADDR")
-
-    assert.Nil(t, err, "getBeaconData should not return error beacon was found")
-
-    key := beaconData{"BEACON_ADDR", "TOKEN", users}
-    assert.Equal(t, key, res, 
-        "getBeaconData should give correct beaconData")
-}
-
-func Test_GetBeaconData_NotFound(t *testing.T) {
-    teardown := setup()
-    defer teardown()
-
-    res, err := getBeaconData("BAD_INST")
-
-    assert.NotNil(t, err, "getBeaconData should forward errors")
-
-    assert.Equal(t, beaconData{}, res, 
-        "getBeaconData should give empty beaconData on error")
 }
 
 func Test_GetBeaconToken_NotFound(t *testing.T) {
@@ -200,133 +112,38 @@ func Test_GetBeaconToken_Valid(t *testing.T) {
         "GetBeaconToken should give corrent token")
 }
 
-func Test_ListBeacons_ValidUser(t *testing.T) {
-    teardown := setup()
-    defer teardown()
+func tryHandleTest(t *testing.T, r *http.Request, m *mux.Router) {
+    defer func() { recover() }()
 
-    keyList := make([]aliases.Alias, 0)
+    w := httptest.NewRecorder()
+    m.ServeHTTP(w, r)
 
-    for i := 1; i <= 2; i++ {
-        beaconList, err := getBeaconsList("USER")
-
-        assert.Nil(t, err, "getBeaconList returned an error")
-
-        assert.Equal(t, keyList, beaconList, 
-            "getBeaconList output differed from key")
-
-        newBeacon := map[string]interface{} {
-            "Address" : fmt.Sprintf("BEACON_ADDR %d", i), 
-            "Token" : "TOKEN", 
-            "Users" : userMap{"USER":true},
-        }
-
-        keyPair := aliases.Alias{"", newBeacon["Address"].(string)}
-
-        keyList = append(keyList, keyPair)
-        beacons.InsertSchema(newBeacon)
-    }
-
-    beaconList, err := getBeaconsList("USER")
-
-    assert.Nil(t, err, "getBeaconList returned an error")
-    assert.Equal(t, keyList, beaconList)
+    // This won't run during a panic(), but we can't panic during a 404
+    assert.NotEqual(t, http.StatusNotFound, w.Code)
 }
 
-func Test_ListBeacons_BadUser(t *testing.T) {
-    teardown := setup()
-    defer teardown()
+func Test_Handle(t *testing.T) {
+    r := mux.NewRouter()
+    Handle(r)
 
-    goodBeacon := map[string]interface{} {
-        "Address" : "BEACON_ADDR 1", 
-        "Token" : "TOKEN", 
-        "Users" : userMap{"GOOD_USER":true},
+    routes := []struct {
+        Method string
+        Endpoint string
+    } {
+        {"PUT",    "/user/TEST"},
+        {"DELETE", "/user/TEST"},
+        {"PUT",    "/token/TEST"},
+        {"POST",   "/create"},
+        {"GET",    "/list"},
+        {"GET",    "/list/TEST"},
+        {"PUT",    "/refresh/TEST"},
     }
 
-    badBeacon := map[string]interface{} {
-        "Address" : "BEACON_ADDR 2", 
-        "Token" : "TOKEN", 
-        "Users" : userMap{"BAD_USER":true},
+    for _, route := range routes {
+        m := route.Method
+        e := route.Endpoint
+
+        req, _ := http.NewRequest(m, e, bytes.NewBuffer([]byte("")))
+        tryHandleTest(t, req, r)
     }
-
-    keyList := []aliases.Alias{aliases.Alias{"", "BEACON_ADDR 1"}}
-
-    beacons.InsertSchema(goodBeacon)
-    beacons.InsertSchema(badBeacon)
-
-    beaconList, err := getBeaconsList("GOOD_USER")
-
-    assert.Nil(t, err, "getBeaconList returned an error")
-    assert.Equal(t, keyList, beaconList)
-}
-
-func Test_ListInstances_ValidUser(t *testing.T) {
-    teardown := setup()
-    defer teardown()
-
-    beacon := map[string]interface{} {
-        "Address" : "BEACON_ADDR", 
-        "Token" : "TOKEN", 
-        "Users" : userMap{"USER":true},
-    }
-
-    beacons.InsertSchema(beacon)
-
-    keyList := make([]instanceData, 0)
-
-    for i := 1; i <= 2; i++ {
-        instanceList, err := getInstancesList("BEACON_ADDR", "USER", false)
-
-        assert.Nil(t, err, "getInstancesList returned an error")
-
-        assert.Equal(t, keyList, instanceList, 
-            "getInstancesList output differed from key")
-
-        newInstance := instanceData {
-            InstanceAddress : fmt.Sprintf("INST_ADDR %d", i), 
-            BeaconAddress : "BEACON_ADDR",
-            Name : fmt.Sprintf("VM %d", i),
-            CanAccessDocker : true,
-        }
-
-        keyList = append(keyList, newInstance)
-
-        instances.InsertSchema(map[string]interface{} {
-            "InstanceAddress" : newInstance.InstanceAddress, 
-            "BeaconAddress" : newInstance.BeaconAddress,
-            "Name" : newInstance.Name,
-            "CanAccessDocker" : newInstance.CanAccessDocker,
-        })
-    }
-
-    instanceList, err := getInstancesList("BEACON_ADDR", "USER", false)
-
-    assert.Nil(t, err, "getInstancesList returned an error")
-    assert.Equal(t, keyList, instanceList)
-}
-
-func Test_ListInstances_BadUser(t *testing.T) {
-    teardown := setup()
-    defer teardown()
-
-    beacon := map[string]interface{} {
-        "Address" : "BEACON_ADDR", 
-        "Token" : "TOKEN", 
-        "Users" : userMap{"GOOD_USER":true},
-    }
-
-    beacons.InsertSchema(beacon)
-
-    instance := map[string]interface{} {
-        "InstanceAddress" : "INST_ADDR", 
-        "Name" : "NAME",
-        "CanAccessDocker" : true,
-        "BeaconAddress" : "BEACON_ADDR", 
-    }
-
-    instances.InsertSchema(instance)
-
-    instanceList, err := getInstancesList("BEACON_ADDR", "BAD_USER", false)
-
-    assert.Nil(t, err, "getInstancesList returned an error")
-    assert.Equal(t, 0, len(instanceList))
 }
